Splitting process_trigger into dataset_scanner and tt_agent. #42: Adding tt_agent.rsync.  Moving proctrigger.main to client.main. Adding Configuration.switch_config.
