+++ analysis_driver/util/__init__.py
@@ -1,2 +1,2 @@
-from .functions import bcbio_prepare_samples, setup_bcbio_run, transfer_output_data
+from .functions import bcbio_prepare_samples, setup_bcbio_run
 from . import fastq_handler
