+++ bin/edingen_analysis_driver.py
@@ -5,6 +5,6 @@ import os.path
 if __name__ == '__main__':
 
     sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-    import analysis_driver.process_trigger
+    import analysis_driver.client
 
-    sys.exit(analysis_driver.process_trigger.main())
+    sys.exit(analysis_driver.client.main())
